Exclure résultats DB de résultats RT
La mise à jour récupère les ID des films récupérés en base de données et
exclut ceux récupérés depuis Rotten Tomatoes ayant la même ID des
résultats à afficher/inclure dans la base.

+++ rt_search_engine/rt.php
@@ -16,6 +16,8 @@
 
 	if(isset($_POST["filmTitle"])){
 		$query=$_POST["filmTitle"];
+			
+		echo '<ul>';
 		
 		//Search for up to five results in DB
 		
@@ -25,51 +27,55 @@
 		
 		$i=0;
 		
+		$arrResult=[
+			"foo" => "bar",
+			"bar" => "foo",
+		];
+		
 		while (($row = mysql_fetch_row($dbSearchResults)) && ($i < 5)){
 			
 			$i++;
 			
-			echo  "<p>" . $i . $row[0] . "</p>";
-		
+			echo  "<li>" . $i . "/" . $row[0] . " : " . $row[2] . "</li>";
+			
+			$arrResult[$row[2]]=$row[0];
+			
 		}
 		
 		//Search Rotten Tomatoes API for extra results up to total of five
 		if($i<5){
-			
-			//EXCLUDE DB RESULTS FROM API RESULTS
 		
-			$rtExtras = 5 - $i;
-		
-			//$query=$_GET["q"];
 			$apikey = '';
 			$q = urlencode($query); // make sure to url encode query parameters
 
 			// construct the query with our apikey and the query we want to make
-			$endpoint = 'http://api.rottentomatoes.com/api/public/v1.0/movies.json?apikey=' . $apikey . '&q=' . $q . '&page_limit=' . $rtExtras;
+			$endpoint = 'http://api.rottentomatoes.com/api/public/v1.0/movies.json?apikey=' . $apikey . '&q=' . $q . '&page_limit=5';
 
 			$search_results = getJsonCurl($endpoint);
-
-			if(isset($_POST["json"])){
-				echo $search_results;
-			}
 
 			// play with the data!
 			$movies = $search_results->movies;
 			
-			echo '<ul>';
-			
-			foreach ($movies as $movie) {
-			
-				echo '<li>';
-				//<a href="http://api.rottentomatoes.com/api/public/v1.0/movies/' . $movie->id . '.json?apikey=' . $apikey . '">' . $movie->title . " (" . $movie->year . ")</a><br>";
-				$details = getDetailMovie($movie->id);
-				echo '</li>';
+			foreach ($movies as $movie){
+				
+				//EXCLUDE DB RESULTS FROM API RESULTS
+				//Problem undefined/uninitialized string offset because array is null
+				
+				if(!array_key_exists($movie->id, $arrResult) && $i < 5){
+				
+					$i++;
+				
+					echo '<li>' . $i ."/";
+					$details = getDetailMovie($movie->id);
+					echo '</li>';
+				
+				}
 			
 			}
-			
-			echo '</ul>';
 		
 		}
+			
+		echo '</ul>';
 		
 	}
 
@@ -87,8 +93,7 @@
 		// remember to close the curl session once we are finished retrieveing the data
 		curl_close($session);
 		
-		/*
-		// decode the json data to make it easier to parse the php
+		/*decode the json data to make it easier to parse the php
 		echo $data;
 		*/
 		
@@ -97,8 +102,9 @@
 		if ($search_results === NULL) die('Error parsing json');
 		
 		return $search_results;
-	}
 		
+	}
+	
 	function getDetailMovie($id){
 	
 		global $apikey;
@@ -132,7 +138,7 @@
 			}
 			
 		}
-		
+	
 		//Get DIRECTOR NAME and send to DB
 		if(isset($result->abridged_directors)){
 			$directors = $result->abridged_directors;
@@ -246,7 +252,7 @@
 		}
 		
 		//Print film details
-		echo '<p><a href="http://api.rottentomatoes.com/api/public/v1.0/movies/' . $result->id . '.json?apikey=' . $apikey . '">' . $title;
+		echo '<a href="http://api.rottentomatoes.com/api/public/v1.0/movies/' . $result->id . '.json?apikey=' . $apikey . '">' . $title;
 		
 		if(isset($result->year) && ! $year==null){
 			echo " (" . $year . "), ";
@@ -279,8 +285,6 @@
 				echo $genre . ", ";
 			}
 		}
-		
-		echo "</p>";
 	
 	}
